requirements.txt
paho-mqtt
python-kasa
python-dotenv
asyncio
pandas
plotly